refactor: extract gateway admin links
